European Beech
The asset was added from the Fab Marketplace using a free personal license. However, when it's moved to a different folder (e.g., the "Fab" folder), it seems to stop working due to missing links between associated files. Using the advanced copy option also doesn't help, as it causes Unreal Engine to crash.